HBSD: Add GO_PIE knob for per-port PIE opt-in
This enables building a golang-derived port as a Position-Independent
Executable (PIE). It's currently opt-in, meaning each port needs to
define GO_PIE in its Makefile.

With time, I hope we can make this feature opt-out rather than opt-in
like we do for normal PIE. An exp-run with opt-out functionality will
need to be run, and functionality tuned appropriately.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX
github-issue:	#159